Fix Cell Balancing (#1309)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->
In the extreme case of an over-charged pack, cell-balancing cannot be
performed to bring the pack back into spec. This PR adjusts the
MAX_CELL_VOLTAGE threshold depending on if balancing is enabled or not,
allowing a slightly over-charged pack to be brought back below 4.2v per
cell through discharging.

Also fixes bug in the isoSPI state machinecaused by recent change with
OWC.

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

Tested on pack

### Resolved Tickets
<!-- Link any tickets that this PR resolves. -->