Fix inverter on state (#1314)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

Currently, "inverter on" state is not being respected. This means the
inverters are not actually given time to boot up before HV is applied.
This is probably fine, since we've been running this for a while, but we
should still fix.

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

Just unit testing ATM.

---------

Co-authored-by: Edwin <20777515+Lucien950@users.noreply.github.com>