USB Drivers (#1404)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->
added a hw_usb layer, which contains CDC class USB calls for
transmitting and receiving function calls.
initialized a queue in the cdc_recieve function to place the information
received under the hood into the queue one byte at a time
### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->
tested the receive by sending a message through python client
tested the transmit function by sending a message and receiving it on
the python end
tested hot-plugging the USB device and seeing if the status is changing 

### Resolved Tickets
<!-- Link any tickets that this PR resolves. -->

---------

Co-authored-by: Liam Ilan <snowboardsheep@gmail.com>