Kevin/bms/imd decoding fix (#1424)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->
Changed the trigger source for tim1 from ti1fp2 to ti1fp1 since we
switched the channels 1 and 2 functionality from thruna to quadruna.
Also changed the user constants for the tim1 frequency to the value of
the clk config.

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->
Tested with function generator and oscilloscope to confirm readings were
accurate.
### Resolved Tickets
<!-- Link any tickets that this PR resolves. -->
FIRM-221