Async I2C driver (#1440)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

Changed our `hw_i2c` driver such that it yields the currently running
FreeRTOS task while a transaction is in progress. This means the CPU is
free to work on other stuff (like logging, CAN, etc.) while the
peripheral is handling the transaction.

The process is something like this:
1. Start an I2C transaction with `hw_i2c_*`, an interrupt mode, so an
interrupt will get fired when completed.
2. Yield the current task with `ulTaskNotifyTake`. The scheduler will
switch to work on something else.
3. When the transaction is completed, an ISR runs which calls
`vTaskNotifyGiveFromISR`. This sends a notification to the task that
started the interrupt, which makes it ready to run.
4. The scheduler can transition to the task that started the I2C
transaction and continue running it.

What's cool about this is we get the benefits of using I2C in interrupt
mode (CPU can do other stuff while the peripheral is cooking) without
the downsides (configuring the callbacks yourself is annoying and error
prone).

## Comparison

VC 100Hz task before this change:


![image](https://github.com/user-attachments/assets/33409b5e-cbf7-4bd3-891e-5411163ce8bc)

VC 100Hz task after this change:


![image](https://github.com/user-attachments/assets/ed366e10-1e42-44fe-b7a9-0245bcdd9ae6)

Notice how the total time to run the 100Hz task is approximately the
same (4.5ms), but the CPU has way more free time after. This is because
the CPU yields during I2C transactions with the IMU chip. This free time
it could spend dealing with CAN, running vehicle dynamics, etc. etc.

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

- [x] Tested transacting with the IMU on the VC

### Resolved Tickets
<!-- Link any tickets that this PR resolves. -->