Cortex-M7 compiler flag fix (#1432)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

Looks like I made a mistake when setting up the compiler flags for
Cortex-M7s and we've been passing `-mcpu=cortex-m4`. They both support
the same instruction set (which I believe this flag is responsible for
configuring) which is probably why we haven't seen any issues. However,
we should definitely correct this regardless. There are differences in
the final binaries so this definitely does something...

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

- [x] Builds pass.
- [x] Sanity test on hardware.