Fix CAN logging bugs (#1508)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

Fix a few bugs in CAN logging that were causing logging to fail
silently. Also reworked some of the error handling so it should be more
robust, and be more clear what's going wrong (due to asserts failing or
RTT logs).

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

- [x] Tested the bug fixes on the car but made some changes since then
that I'll need to re-verify