Async UART driver (#1453)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

Changed our hw_uart driver such that it yields the currently running
FreeRTOS task while a transaction is in progress. This means the CPU is
free to work on other stuff (like logging, CAN, etc.) while the
peripheral is handling the transaction. But also still provides an API
for callback-style transactions.

Same treatment to UART as I2C in
https://github.com/UBCFormulaElectric/Consolidated-Firmware/pull/1440.


### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

- [x] Test on SBG
- [x] Test on old Chimera

---------

Co-authored-by: Edwin <20777515+Lucien950@users.noreply.github.com>