Reset before booting (#1613)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

Currently the bootloader, when jumping to the app, de-inits everything
it inits and then jumps to the app. This is a bit scuffed since then we
need to manually deinit everything - this is error prone and we were
seeing some problems with the clocks.

Now, when the bootloader wants to start the app, it writes to the boot
flag, resets itself, and upon the next bootup (before anything is
initialized) jumps to the app. This is the same mechanism as the app
jumping to the bootloader.

Also switched the bootloader to use software CRC (`app_crc`) since it
needs to check the app's validity before initializing the CRC peripheral
now.

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

Tested booting to app and bootloading new binaries on the FSM.