ASYNC CANTX (#1591)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->
- working on f4
- working on h7

![pre](https://github.com/user-attachments/assets/3d8473db-bb45-4837-aed4-1d06305736bf)
This next image depects Edwin Zheng connecting the CAN bus to the F4.
Note the lack of runtime before the plug in event.

![post](https://github.com/user-attachments/assets/2b7ffe69-bf29-4063-a0bc-e7a3493b598c)


### Resolved Tickets
<!-- Link any tickets that this PR resolves. -->
[FIRM-395](https://ubcformulaelectric.atlassian.net/browse/FIRM-395)


[FIRM-395]:
https://ubcformulaelectric.atlassian.net/browse/FIRM-395?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ