Improved stack monitor (#1532)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

Reworked our stack monitor driver. What happens now is when FreeRTOS
detects a stack overflow in one of the thread stacks, it calls
`vApplicationStackOverflowHook`. This writes to a "context" variable in
the boot request RAM which indicates a stack overflow, and then resets
the MCU. Then, on the next boot cycle, it reads this context variable
and populates the CAN table (as well as the task number which caused the
overflow).

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

Tested on the CRIT.