Run LTC conversions at specified intervals (#1627)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

1. Run LTC conversions at specified intervals (10Hz for cells, 1Hz for
temps, 0.1Hz for status and OWC). This involved splitting the broadcast
into 2 parts: 1 part that does the IO operations and 1 part that does
the parsing of results. Did it like this so IO operations for cells can
be ongoing while parsing OWC results, or other app code can be accessing
LTC results, without waiting on those IO operations.
2. Drop SPI bus speed to 31kHz (may be able to crank it higher, I've
just been testing with this)
3. Extend JSONCAN configs for 10x segments (thank you ChatGPT)
4. Slightly tweaked the OWC results parsing. I think we had 1 or 2
off-by-one errors and the most significant pin needs to be special
cased.
5. Everything else is basically the same, and fixed a few bugs
6. Some fixes for FDCAN with >8 bytes

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

Tested with 6 segments total. This took 84ms so we're getting close to
the 10Hz frequency for cells. It suspect we'll probably have to either
crank the SPI frequency, or drop the cell conversion frequency.