Build system fixes (#1640)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

- Re-enables warnings on STM32 `cubemx/` code, except for specific ones
we choose to disable warnings for. In practice this is only syscalls and
USB. We missed a few bugs when testing since warnings weren't being
thrown on `main.c`. Fixed a few bugs on VCR that I think were the result
of this.
- Disable FDCAN protocol exception (I don't think this is something we
want to use?)
- Disable charger CAN on BMS since it was clashing with FDCAN bus (TODO:
Enable/disable this based on charger connected or something else)
- Change how the CRC32 checksum is calculated. Previously the binary
length was rounded up to the nearest word, since the hardware CRC
operates on entire words. Changed to be over the exact amount of bytes
since software CRC does not have this limitation. This was causing a bug
where padded bytes were reading 0x00 on the microcontroller, but the
image generation script assumed they would be 0xFF. (TODO: Why is this
happening? They should be 0xFF as erased flash cells but maybe the JLink
was padding zeros?)

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

- [x] Tested FDCAN on BMS
- [ ] Tested bootloading on BMS
- [ ] Tested bootloading on FSM