Tracksight Frontend Rewrite (#1752)
Rewrites the tracksight frontend with the updated design as well as
cleaning up the state management for signals, while maintaining
relatively the same interface. Should be easy to add in the graph's
because the state has pretty much the same interface.

# Changes

- Removed shadcn components because a lot of the interactivity is pretty
niche (like the signal selector)
- Rewrote state logic to expose two functions
- `useSignalDataRef(signals: string[])`: Returns an immutable object
that never changes containing references to the signals datas which are
never recreated ensuring the garbage collector doesn't get overrun every
render.